service/src/main/java/com/playhub/game/boggle/manager/dao/converters/DurationToLongConverter.java
package com.playhub.game.boggle.manager.dao.converters;

import jakarta.persistence.AttributeConverter;
import jakarta.persistence.Converter;

import java.time.Duration;

@Converter
public class DurationToLongConverter implements AttributeConverter<Duration, Long> {

    @Override
    public Long convertToDatabaseColumn(Duration attribute) {
        return attribute != null ? attribute.toNanos() : null;
    }

    @Override
    public Duration convertToEntityAttribute(Long dbData) {
        return dbData != null ? Duration.ofNanos(dbData) : null;
    }

}
